# static/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: NW Monitor API
  description: |
    REST API for the Network Monitor System.
    Provides endpoints for device management, monitoring, topology,
    dashboards, alerts, SLA reports, bandwidth, SNMP traps, syslog,
    user management, audit logs, and system administration.
  version: "1.2.0"
  contact:
    name: NW Monitor Admin
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/topology/connection:
    post:
      tags: [Topology]
      summary: Add a topology connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_id, connected_to]
              properties:
                device_id: { type: integer }
                connected_to: { type: integer }
                view_type: { type: string, default: standard }
      responses:
        "201":
          description: Connection created
  /api/topology/connection/{connection_id}:
    delete:
      tags: [Topology]
      summary: Delete a topology connection
      parameters:
        - name: connection_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Connection deleted

  # ---- Sub-Topologies ----
  /api/sub-topologies:
    get:
      tags: [Sub-Topologies]
      summary: List all sub-topologies
      responses:
        "200":
          description: Array of sub-topologies
    post:
      tags: [Sub-Topologies]
      summary: Create a sub-topology
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                device_ids:
                  type: array
                  items: { type: integer }
                connections:
                  type: array
                  items: { type: object }
      responses:
        "201":
          description: Sub-topology created
  /api/sub-topologies/{sub_topo_id}:
    get:
      tags: [Sub-Topologies]
      summary: Get a sub-topology with devices
      parameters:
        - $ref: "#/components/parameters/SubTopoId"
      responses:
        "200":
          description: Sub-topology detail
    put:
      tags: [Sub-Topologies]
      summary: Update a sub-topology
      parameters:
        - $ref: "#/components/parameters/SubTopoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                device_ids:
                  type: array
                  items: { type: integer }
      responses:
        "200":
          description: Sub-topology updated
    delete:
      tags: [Sub-Topologies]
      summary: Delete a sub-topology
      parameters:
        - $ref: "#/components/parameters/SubTopoId"
      responses:
        "200":
          description: Sub-topology deleted
  /api/sub-topologies/upload-bg:
    post:
      tags: [Sub-Topologies]
      summary: Upload background image
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload result with URL

  # ---- Dashboards ----
...
  /api/users:
    get:
      tags: [Users]
      summary: List all users (admin)
      responses:
        "200":
          description: Array of users
    post:
      tags: [Users]
      summary: Create a new user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created
  /api/users/{user_id}:
    put:
      tags: [Users]
      summary: Update a user (admin)
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [admin, operator, viewer] }
                display_name: { type: string }
                email: { type: string }
                is_active: { type: boolean }
                password: { type: string }
      responses:
        "200":
          description: User updated
    delete:
      tags: [Users]
      summary: Delete a user (admin)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User deleted
...
  /api/users/me/password:
    put:
      tags: [Users]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string }
      responses:
        "200":
          description: Password changed

  # ---- Audit Log ----
...
components:
  parameters:
...
    SubTopoId:
      name: sub_topo_id
      in: path
      required: true
      schema: { type: integer }
    UserId:
      name: user_id
      in: path
      required: true
      schema: { type: integer }
  schemas:
...
    UserCreate:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        role: { type: string, enum: [admin, operator, viewer], default: viewer }
        display_name: { type: string }
        email: { type: string, format: email }